find_library_file: port from doctest to normal test and fix for cygwin
Starting with 11d33880d89729ade7e8ee7 UnixCCompiler got some improved support
for cygwin which made the doctest fail.

The doctest is hard to read as is, and adding more platform specific exceptions
wouldn't help, so just convert to a normal test and special case cygwin to make
the test pass there again.